Stworzenie kontrolera paginacji
- oznaczenie BookController.class jako @Depracated
- usunięcie nie potrzebnego kodu z kontrolera książki
