+++ src/main/java/edition/academy/seventh/controller/BookController.java
@@ -1,30 +1,26 @@
 package edition.academy.seventh.controller;
 
-import edition.academy.seventh.database.connector.ConnectorFactory;
 import edition.academy.seventh.database.model.BookDto;
-import edition.academy.seventh.display.*;
 import edition.academy.seventh.service.BookService;
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import static edition.academy.seventh.database.connector.DatabaseTypes.POSTGRESQL;
+import java.util.List;
 
 /** @author Kamil Rojek */
 @RestController
 @CrossOrigin(origins = "*", maxAge = 3600)
+@Deprecated
 class BookController {
   private BookService bookService;
-  private Pagination<BookDto> pagination;
 
   @Autowired
   public BookController(BookService bookService) {
     this.bookService = bookService;
-    this.pagination = new LazyPagination(ConnectorFactory.of(POSTGRESQL));
-//    this.pagination = new EagerPagination(bookService);
   }
 
   /**
@@ -35,36 +31,5 @@ class BookController {
   @GetMapping("/books")
   public ResponseEntity<List<BookDto>> getBooks() {
     return new ResponseEntity<>(bookService.getBooksFromDatabase(), HttpStatus.OK);
-  }
-
-  @GetMapping("/books/pagination")
-  public ResponseEntity<List<BookDto>> getBooksWithPagination() {
-    List<BookDto> paginationUsingSql = pagination.currentPage();
-    return new ResponseEntity<>(paginationUsingSql, HttpStatus.OK);
-  }
-
-  @RequestMapping("/books/pagination/size")
-  public ResponseEntity<List<BookDto>> changePaginationSize(
-      @RequestParam("value") PaginationSize size) {
-    List<BookDto> bookDtos = pagination.changePaginationSize(size);
-    return new ResponseEntity<>(bookDtos, HttpStatus.OK);
-  }
-
-  @GetMapping("/books/pagination/next")
-  public ResponseEntity<List<BookDto>> next() {
-    List<BookDto> paginationUsingSql = pagination.nextPage();
-    return new ResponseEntity<>(paginationUsingSql, HttpStatus.OK);
-  }
-
-  @GetMapping("/books/pagination/previous")
-  public ResponseEntity<List<BookDto>> previous() {
-    List<BookDto> paginationUsingSql = pagination.previousPage();
-    return new ResponseEntity<>(paginationUsingSql, HttpStatus.OK);
-  }
-
-  @RequestMapping("/books/pagination/filter")
-  public ResponseEntity<List<BookDto>> setFilter(@RequestParam("type") Filter filter) {
-    List<BookDto> bookDtos = pagination.changeFilter(filter);
-    return new ResponseEntity<>(bookDtos, HttpStatus.OK);
   }
 }
